ai-project/backend/requirements.txt
MarkupSafe-3.0.3 
annotated-doc-0.0.4 
annotated-types-0.7.0 
anyio-4.13.0 
certifi-2026.2.25 
charset_normalizer-3.4.7 
click-8.3.2 
colorama-0.4.6 
contourpy-1.3.3 
cycler-0.12.1 
fastapi-0.136.0 
filelock-3.28.0 
fonttools-4.62.1 
fsspec-2026.3.0 
h11-0.16.0 
idna-3.11 
jinja2-3.1.6 
kiwisolver-1.5.0 
matplotlib-3.10.8 
mpmath-1.3.0 
networkx-3.6.1 
numpy-2.4.4 
opencv-python-4.13.0.92 
packaging-26.1 
pillow-12.2.0 
polars-1.39.3 
polars-runtime-32-1.39.3 
psutil-7.2.2 
pydantic-2.13.2 
pydantic-core-2.46.2 
pyparsing-3.3.2 
python-dateutil-2.9.0.post0 
pyyaml-6.0.3 
requests-2.33.1 
scipy-1.17.1 
setuptools-81.0.0 
six-1.17.0 
starlette-1.0.0 
sympy-1.14.0 
torch-2.11.0 
torchvision-0.26.0 
typing-extensions-4.15.0 
typing-inspection-0.4.2 
ultralytics-8.4.38 
ultralytics-thop-2.0.18 
urllib3-2.6.3 
uvicorn-0.44.0